Adds GitHub action publish.yaml
This adds a `.github/workflows/publish.yaml` for publishing pushes to the `main` branch automatically.

For this to work though I think you need to run `quarto publish` locally first, selecting `GitHub Pages` and letting it
create and push the `gh-pages` branch. The settings on GitHub then need changing, there are two possible places that you
need to check...

1. _Settings > Actions > General > Workflow permissions_ should have _Read and write permissions_ selected. This allows
  the workflow to read from the `main` branch and write to `gh-pages` branch.
2. _Settings > Pages_ probably need changing the _Source_ should be _Deploy from a branch_ and the _Branch_ should be
`gh-pages` and the directory should be `/ (root)`.